% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_mat.R
\name{cbind_list}
\alias{cbind_list}
\title{Takes a list of vectors and cbinds them to form a data frame.}
\usage{
cbind_list(lst)
}
\arguments{
\item{lst}{A list}
}
\description{
If the vectors are of different lengths, then it will pad the
smaller vectors with \code{NA}'s. Any non-vectors will be coerced
to be vectors.
}
\examples{
lst <- list(1:3, 1:4)
cbind_list(lst)

lst <- append(lst, c("a"))
cbind_list(lst)
}
\author{
David Gerard
}

