% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_mat.R
\name{pad_length}
\alias{pad_length}
\title{Pad the ends of a vector with NAs so that it is a specified length.}
\usage{
pad_length(x, n)
}
\arguments{
\item{x}{A vector of length less than \code{n}.}

\item{n}{A positive integer. The desired length of a vector.}
}
\description{
Pad the ends of a vector with NAs so that it is a specified length.
}
\author{
David Gerard
}

